backend/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
chromadb==0.5.5
google-generativeai==0.7.2
python-dotenv==1.0.1
